[savedObjects] change the number mappings to be 'long'

+++ src/kibana/plugins/discover/saved_searches/_saved_search.js
@@ -19,10 +19,10 @@ define(function (require) {
         mapping: {
           title: 'string',
           description: 'string',
-          hits: 'integer',
+          hits: 'long',
           columns: 'string',
           sort: 'string',
-          version: 'integer'
+          version: 'long'
         },
 
         defaults: {
